pyproject.toml
[tool.poetry]
name = "nowplayingbot"
version = "0.1.0"
description = ""
authors = ["es3n1n <me@es3n.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.7.0"
httpx = {extras = ["http2"], version = "^0.27.0"}
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
loguru = "^0.7.2"
pydantic-settings = "^2.3.3"
filelock = "^3.15.1"
async-lru = "^2.0.4"
scdl = {git = "https://github.com/flyingrub/scdl", rev = "42fbc5c23e3331059b05657f0819ac4835d6b83b"}
pycryptodome = "^3.20.0"
pyjwt = "^2.8.0"
yandex-music = "^2.2.0"
websockets = "^12.0"
yt-dlp = "2024.6.17.232743.dev0"
asyncpg = "^0.29.0"
cryptography = "^42.0.8"
apple-music-python = "^1.0.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.10.0"
isort = "^5.13.2"
flake8-quotes = "^3.4.0"
flake8-bugbear = "^24.4.26"
flake8-print = "^5.0.0"
asyncpg-stubs = "^0.29.1"
wemake-python-styleguide = "^0.19.2"

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
lines_after_imports = 2
combine_as_imports = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
